requirements.txt
azure-identity
requests
pandas
python-dotenv
autogenstudio
powerbiclient
openai
jsonschema
autogen
